chore(shared): clean up Rspack and server types
