+++ .vscode/settings.json
@@ -3,6 +3,7 @@
   "search.exclude": {
     "**/.git": true,
     "**/dist": true,
+    "**/dist-types": true,
     "**/coverage": true,
     "**/compiled": true,
     "**/doc_build": true,
