+++ packages/core/src/server/devMiddleware.ts
@@ -1,12 +1,14 @@
+import type { IncomingMessage, ServerResponse } from 'node:http';
 import {
   type CompilerTapFn,
   type DevConfig,
-  type DevMiddleware,
+  NextFunction,
   applyToCompiler,
   isClientCompiler,
   isNodeCompiler,
 } from '@rsbuild/shared';
 import type { Compiler, MultiCompiler } from '@rspack/core';
+import { DevMiddlewareOptions } from '../provider/createCompiler';
 
 type ServerCallbacks = {
   onInvalid: () => void;
@@ -64,6 +66,24 @@ function applyHMREntry({
     }).apply(compiler);
   }
 }
+
+export type Middleware = (
+  req: IncomingMessage,
+  res: ServerResponse,
+  next: NextFunction,
+) => Promise<void>;
+
+export type DevMiddlewareAPI = Middleware & {
+  close: (callback: (err: Error | null | undefined) => void) => any;
+};
+
+/**
+ * The rsbuild/server do nothing about compiler, the devMiddleware need do such things to ensure dev works well:
+ * - Call compiler.watch （normally did by webpack-dev-middleware）.
+ * - Inject the hmr client path into page （the hmr client rsbuild/server already provide）.
+ * - Notify server when compiler hooks are triggered.
+ */
+export type DevMiddleware = (options: DevMiddlewareOptions) => DevMiddlewareAPI;
 
 export const getDevMiddleware = async (
   multiCompiler: Compiler | MultiCompiler,
