+++ packages/core/src/types.ts
@@ -15,6 +15,16 @@ declare module '@rspack/core' {
   }
 }
 
+export type RspackSourceMap = {
+  version: number;
+  sources: string[];
+  mappings: string;
+  file?: string;
+  sourceRoot?: string;
+  sourcesContent?: string[];
+  names?: string[];
+};
+
 export type { RsbuildPlugin, RsbuildPlugins, RsbuildPluginAPI };
 
 /** The inner context. */
